+++ storjscandb/dbx/storjscandb.dbx.pgx.sql
@@ -9,7 +9,7 @@ CREATE TABLE block_headers (
 );
 CREATE TABLE token_prices (
 	interval_start timestamp with time zone NOT NULL,
-	price double precision NOT NULL,
+	price bigint NOT NULL,
 	PRIMARY KEY ( interval_start )
 );
 CREATE TABLE wallets (
